For each calstis module, the function that reads the command-line arguments now checks for "--version", and if present, the function prints the calstis version number (without the date) and exits with status 0.  Function PrVersion to extract the number and print it was added to lib/prinfo.c. See Trac #734.
The following changes are for making use of the IMPHTTAB and functions
InitPhotPar, GetPhotTab, and FreePhotPar.

Changes to cs0/sciflags.c and cs0/wavflags.c:

For imaging data, check for IMPHTTAB instead of PHOTTAB.

Changes to cs1/calstis1.h:

filtcorr, tdscorr, apertab and tdstab were removed from the StisInfo1
struct.  The PhotInfo struct was deleted.

Changes to cs1/calstis1.c:

Function CalStis1 was modified to not set sts.filtcorr or sts.tdscorr.
Function StisInit1 was modified to not call InitRefTab for apertab or for
tdstab.

Changes to cs1/do2d.c:

The call to PhotMode was moved to a point just before the call to doPhot.
In PhotMsg, keyword phottab was changed to imphttab, and info is no longer
printed about apertab or tdstab.

Changes to cs1/dophot.c:

Functions InitPhotPar, GetPhotTab, and FreePhotPar are called, instead of
explicitly getting the throughput info and calling c_phopar.

Changes to cs1/getflags1.c:

Function checkPhot was modified to check for keyword IMPHTTAB instead of
PHOTTAB, and keywords APERTAB and TDSTAB are ignored.  sts->filtcorr and
sts->tdscorr are no longer set.

Changes to cs1/history1.c:

Function photHistory was modified to not check filtcorr or tdscorr.

Changes to cs1/photmode.c:

The Modified Julian Date was added to the photmode string (MJD#<date>).

Changes to cs2/cs2.c:

cs2_command returns a status code, but it wasn't being checked.  A test
on the value returned by cs2_command was added; if it's not zero, exit
is called with that value of status.

The following changes make this version consistent with the iraf-dependent
version of calstis 2.36.

PR 64987:  CALSTIS should use reference temp and slope from CCD dark file
PR 65624:  Add another term to the dispersion relation in CALSTIS.
PR 66617:  Rewrite calstis function for inverting the dispersion relation.
PR 67122:  Improve the a4corr correction for echelle wavecal processing
           in calstis.
PR 68052:  Calstis failure when an imset is missing (exptime was 0).

The following files were modified:
    include/stis.h
    include/stisversion.h
    cs1/dodark.c
    cs4/calstis4.h
    cs4/echshift.c
    cs4/getdisp4.c
    cs4/getkeyinfo4.c
    cs4/maketemplate.c
    cs4/wavecal.c
    cs4/xc4.c
    cs4/xcwave.c
    cs6/commline.c
    cs6/wave.c
    cs7/x2dcorr7.c
    lib/checkimsetok.c

The following files were deleted:
    cs1/getapthr1.c
    cs1/getphot1.c
    cs1/gettds1.c
    cs6/evaldisp6.c
    cs7/evaldisp.c

The following files were added:
    cs7/evaldisp7.c
    lib/evaldisp.c

CALSTIS was using hard-coded parameters for the reference temperature
and slope in order to correct the temperature dependence of the CCD
dark rate.  Analysis of SMOV data shows that these values have changed.
Rather than including time-dependent parameters in calstis, these are
now included as header keywords in the dark reference files, and
calstis has been modified to read and use these keyword values.

A ninth term has been added to the dispersion relation.  This is
proportional to m**2, where m is the spectral order.  If wl is the
wavelength in Angstroms, the dispersion relation is:
    pixel = coeff[0] +
            coeff[1] * m * wl +
            coeff[2] * m**2 * wl**2 +
            coeff[3] * m +
            coeff[4] * wl +
            coeff[5] * m**2 * wl +
            coeff[6] * m * wl**2 +
            coeff[7] * m**3 * wl**3 +
            coeff[8] * m**2

The a4corr correction is a modification to the echelle dispersion
coefficients to account for a skew that results from a displacement of the
image in the cross-dispersion direction.  This displacement is determined
from shifta2 (they are the same except for zero point).  However, shifta2
is not known at the beginning of wavecal processing, because it's one of
the parameters that are determined by wavecal processing.  Previously,
keyword moffset2 was used as an estimate of shifta2, but for recent data
this keyword is not updated.  The current solution is that part of the
wavecal processing for echelle data is now done twice, the first time with
a4corr set to zero, just to get the value of shifta2, which is then used
to compute the a4corr correction, and the second time with a4corr included.

When checking whether an image set exists or not, calstis opens the
SCI,EXTVER extension using c_tbtopn, a table I/O function.  If this fails,
c_tbtclo was called to close the extension (usually an image), and -1 was
returned as a flag.  The call to c_tbtclo failed because the file handle
was not assigned a value; it is also unnecessary to call c_tbtclo because
memory had already been freed.  This call to c_tbtclo has been deleted.


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/hstcal/trunk@13391 fe389314-cf27-0410-b35b-8c050e845b92