The buffer sizes for input file lists for cs0 and cs1 are now allocated based on the sizes of command-line strings (or 1), rather than using fixed sizes.
The image name template functions will now copy a file name to the internal
list of names regardless of whether the file exists or not.  Also, if the
user specified an expression in brackets at the end of a file name, that
expression is preserved (although this would not normally work with, say,
hstio).

See Trac #847, #848; PR 70269, 70330.


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/hstcal/trunk@14467 fe389314-cf27-0410-b35b-8c050e845b92