During cosmic-ray rejection, there were two input images that were opened but never closed.  This caused a problem when calstis2 tried to open the file containing them read/write in order to update the DQ extension. See Trac #899.
In calstis2 there was a hard-coded limit of 120 on the number of input image
sets.  This number was used for the size of several arrays.  These arrays
are now allocated with a length equal to the actual number of input image
sets.  The number that are actually used can be smaller, e.g. if one or more
image sets had zero exposure time.
See Trac #900.


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/hstcal/trunk@16932 fe389314-cf27-0410-b35b-8c050e845b92