In function X1DBack in cs6/background.c, the loop over background pixels in a column used floating-point loop index and limit, and the actual number of background pixels that were included could vary due to roundoff. This was modified to use integer index and limit.  See Trac #1040.
The code to select the row to use from a post-SM4 TDCTAB failed when
the best row was the first in the table.  Variable read_this_row in
cs1/gettdccorr.c was initialized to -1 (intentionally not a valid row
number), but it was never reset to an actual row number if the exposure
start time was between the DATE0 values for the first and second rows.
See Trac #916.


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/hstcal/trunk@26704 fe389314-cf27-0410-b35b-8c050e845b92