Fixes #1037, PR 75396 PSF sampling bug.
In the section for echelle scattered light correction, function GetPSF reads
the PSF image from a reference table (_tel.fits).  If the aperture that was
used for the exposure is smaller than the PSF image, GetPSF then truncates
the PSF image to include only the portion that would have gotten through the
aperture.  The PSF image is oversampled quite a bit, and it needs to be binned
down to match the STIS image.  The number of STIS image pixels for the PSF
(truncated by the aperture) is found from the aperture size and plate scale,
but this computation did not take into consideration the possibility that
the aperture may actually be larger than the PSF image.  This has been fixed
by modifying the two lines for computing ms and ml.



git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/hstcal/trunk@29967 fe389314-cf27-0410-b35b-8c050e845b92