ACS CTE gen2
This incorporates into calacs the generation 2 CTE algorithm used in calwf3.
The implementation of these algorithms are not identical though they have been
tested to yield identical results.

This change introduces three new command line options to acscte.e and calacs.e.
These are:

--nthreads <N> : specifies the number of threads to use for the CTE correction
--pctetab <filename> : use 'filename' as the PCTETAB cte reference file instead
               of that specified by PCTETAB in the image header
--ctegen <1|2> : indicates intent of which cte algorithm to use, gen1 or gen2.
             This is a safe guard against incorrect use when the code
         determines which to use dependent on which PCTETAB generation
         was specified. See the below description.

The code looks for CTE_NAME in the specified PCTETAB header, if it is missing it
assumes gen1 cte. If it is = "pixelCTE 2017", gen2 is used. Defaults to gen1
for anything else. The cmd line opt --ctegen <1|2> can be used to give intent
and safe guard, failing if incorrect PCTETAB given.

Signed-off-by: James Noss <jnoss@stsci.edu>