All BIAS files are processed by CALACS to use only read noise for ERR array
Fixes #277.

All BIAS files are processed by CALACS to use only read noise (vs Poisson)
for the computation of the ERR array as the date discriminant has been
removed. Also, when using ACSREJ standalone, the user now must use the
"readnoise_only" flag (formerly newbias) to obtain the same results as CALACS
for BIAS images.  When updating ACSREJ for standalone use, improved the
syntax or usage message which addresses a portion of #278 for ACSREJ only.