Issue 326 darktime updates (#467)
* Modified reading of CCDTAB

* Removed hard-coded darktime scaling value and read new post-flashed
and unflashed columns from an updated CCDTAB to use for the offset to
the DARKTIME keyword value for appropriate date and supported subarrary
criteria.  The DARKTIME keyword value is the default scaling factor for
the superdarks, with the offset being an additive correction.

* Update to ensure the supported subaperture array is complete.

* Added a new PyTest to accommodate the updated darktime correction.
Full-frame data is categorized as pre- and post-SM4, and post-flashed or
unflashed.  Subarray data is categorized as pre- and post-Cycle24, and
post-flashed or unflashed.  The post-Cycle24 data must also be a
"supported" subarray to have the offset adjustment applied.

* Added curley braces to all "if" blocks in dodark.c, and removed some
deprecated code.  Set the full frame CTE tests in the new Pytest,
test_wfc_darktime.py, to SKIP.