HLA-1136: Flag full-well saturated pixels in the overscan regions in DQICORR (#597)
* Updates to reconcile the old and new algorithms for flagging full-well saturated pixels.
The new algorithm uses a saturation image which is applied after BLECORR/BIASCORR, and
the image is already trimmed of the overscan regions. The old algorithm used a single
scalar value applied to the ENTIRE readout (science data and overscan).  Added flagging
in DQICORR of the overscan regions when the saturation image of the new algorithm is to
be applied later in the processing.

* Removed duplicate line and old comment with timestamp