Only commit change to address garbled text and reporting of PEDIGREE and DESCRIP. (#598)
Fixed the bug where the structure associated with the SPOTTAB variables was not
initialized.  Also, fixed the issue where the PEDIGREE and DESCRIP keywords were
delibrately being read from the wrong HDU.  Only in the case of tabular reference
files were the mentioned keywords being read from the wrong HDU.  Fixed so the
keywords are now being read from the primary HDU and properly logged to the trailer
file.